#21 Refactor: refactor admin controller package
